backend/requirements-render.txt
fastapi==0.78.0
uvicorn==0.17.6
python-multipart==0.0.5
python-docx==0.8.11
PyPDF2==2.11.1
python-dotenv==0.19.2
requests==2.27.1
pydantic==1.8.2
gunicorn==20.1.0 